Create `Hexagon.placed()` method
For generating new version of the graph without the nodes occluded
by the given placement. Required a `deepcopy` method which is also
implemented. This method requires some changing to the constructor
which only minimally change behavior--namely, that `Hexagon` can
now be instantiated with an arbitrary dictionary, although it still
defaults to the r=3 Hexagon if nothing is passed.