Implement `Hexagon.count_better()`
This uses a much better recursive relationship which removes all
counting duplicatation. The counting problem is split into two
parts: counting all tilings where an arbitrarily selected triangle
is the center of a tile placement, and counting all tilings where
the selected triangle is neccessarily NOT the center of a tile
placement, i.e., where the triangle is occulded by a tile placement
about one of its neighbors. Additionally, this removes all concern
of tiling "ordering."