Change graph access pattern
Instead of actually modifying the "graph" data structure itself,
we simple "hide" the removed nodes. This is accomplished by
providing a number of overrides such that the interface still
behaves (largely) the same, but `Hexagon.remove_node()` is replaced
with `Hexagon.hide_node()`. Methods like `.keys()`, `.items()`,
and functionality such as `len()`, indexing, containment checking,
behave as in a `dict`. To be more complete, we could go even
further and override the deletion of elements which would use a
lot of the logic from `.hide()`, but eh.