Use bitmask for `self.deleted`
Unfortunately, this doesn't seem to make the code any faster, but
after adding the three short helper functions, `delete()`,
`undelete()`, and `deleted()`, the rest of the code is a bit
cleaner. Also, bitmasks are cool. Also, it might enable as to do a
sort of tree of "total population" tracking for efficiently finding
non-deleted nodes and also checking if all nodes are deleted.