Maintain a "total population" tree
For every node in the "total population tree," if it's a leaf node,
then it describes the deleted status, just as before. If it's an
interior node, it is simply the logical AND of it's children. Using
this, we can find a non-deleted node in log(N) steps instead of N
steps. This is actually just implemented as an additional 64 (more
significant) bits on the `self.deleted` bitmask.

child / 2 + 64 is a given node's parent.

2 * (parent - 64) and 2 * (parent - 64) + 1 are a given node's
children.

As it is, this approach makes the program slower, but I suspect
this is because of the maintence costs of the tree and in spite of
the actual search (recursion termination condition) being faster.
It may still be worth it to try to speed up the maintenance.